Made the util list function a component export of the Scripture component.
