+++ packages/apolloschurchapp/src/ui/Scripture/index.js
@@ -4,14 +4,20 @@ import { get } from 'lodash';
 import PropTypes from 'prop-types';
 import { View } from 'react-native';
 
-// import HTMLView from 'apolloschurchapp/src/ui/HTMLView';
 import Spacer from 'apolloschurchapp/src/ui/Spacer';
 
 import Item from './Item';
 import getScripture from './getScripture';
 
-// const copyright =
-//   '<h6>Scripture taken from The Holy Bible, English Standard Version. Copyright &copy;2001 by <a href="http://www.crosswaybibles.org">Crossway Bibles</a>, a publishing ministry of Good News Publishers. Used by permission. All rights reserved. Text provided by the <a href="http://www.gnpcb.org/esv/share/services/">Crossway Bibles Web Service</a><h6>';
+export const ScriptureList = ({ scripture, commas = true }) => {
+    if (!scripture && commas) return '';
+    if (!scripture) return [];
+
+    const combo = scripture.map(({ reference }) => `${reference}`);
+
+    if (commas) return combo.join(', ');
+    return combo;
+}
 
 const Scripture = ({ references = [] }) => (
   <View>
@@ -27,7 +33,6 @@ const Scripture = ({ references = [] }) => (
       </Query>
     ))}
     <Spacer byHeight />
-    {/* <HTMLView>{copyright}</HTMLView> */}
   </View>
 );
