--- packages/apolloschurchapp/src/utils/scriptures/index.js
@@ -1,5 +1,0 @@
-import list from './list';
-
-export default {
-  list,
-};
